pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.69.0",
    "nltk==3.9.1",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.2",
    "pillow>=11.3.0",
    "sentence-transformers>=5.2.3",
]
